third_party/blink/web_tests/http/tests/loading/empty-subframe-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
frame "f1" - BeginNavigation request to 'about:blank', http method GET
frame "f1" - DidStartNavigation
frame "f1" - ReadyToCommitNavigation
frame "f1" - didCommitLoadForFrame
frame "f1" - didFinishDocumentLoadForFrame
frame "f1" - didHandleOnloadEventsForFrame
frame "f1" - didFinishLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a test of load callbacks. It is only useful inside the regression test tool.
